[Platform][Refactor] Extract func `get_default_attn_backend` to `Platform` (#10358)
