+++ vllm/platforms/openvino.py
@@ -3,13 +3,19 @@ import torch
 import vllm.envs as envs
 from vllm.logger import init_logger
 
-from .interface import Platform, PlatformEnum
+from .interface import Platform, PlatformEnum, _Backend
 
 logger = init_logger(__name__)
 
 
 class OpenVinoPlatform(Platform):
     _enum = PlatformEnum.OPENVINO
+
+    @classmethod
+    def get_default_attn_backend(cls, selected_backend: _Backend) -> _Backend:
+        if selected_backend != _Backend.OPENVINO:
+            logger.info("Cannot use %s backend on OpenVINO.", selected_backend)
+        return _Backend.OPENVINO
 
     @classmethod
     def get_device_name(self, device_id: int = 0) -> str:
